FIX-#2054: modin/utils.py created
