+++ modin/pandas/dataframe.py
@@ -33,11 +33,11 @@ from typing import Tuple, Union
 import warnings
 
 from modin.error_message import ErrorMessage
+from modin.utils import _inherit_docstrings
 from .utils import (
     from_pandas,
     from_non_pandas,
     to_pandas,
-    _inherit_docstrings,
     hashable,
 )
 from .iterator import PartitionIterator
