+++ modin/pandas/utils.py
@@ -11,8 +11,6 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
-import pandas
-
 
 def from_non_pandas(df, index, columns, dtype):
     from modin.data_management.dispatcher import EngineDispatcher
@@ -68,84 +66,6 @@ def to_pandas(modin_obj):
         A new pandas DataFrame or Series.
     """
     return modin_obj._to_pandas()
-
-
-def _inherit_docstrings(parent, excluded=[]):
-    """Creates a decorator which overwrites a decorated class' __doc__
-    attribute with parent's __doc__ attribute. Also overwrites __doc__ of
-    methods and properties defined in the class with the __doc__ of matching
-    methods and properties in parent.
-
-    Args:
-        parent (object): Class from which the decorated class inherits __doc__.
-        excluded (list): List of parent objects from which the class does not
-            inherit docstrings.
-
-    Returns:
-        function: decorator which replaces the decorated class' documentation
-            parent's documentation.
-    """
-
-    def decorator(cls):
-        if parent not in excluded:
-            cls.__doc__ = parent.__doc__
-        for attr, obj in cls.__dict__.items():
-            parent_obj = getattr(parent, attr, None)
-            if parent_obj in excluded or (
-                not callable(parent_obj) and not isinstance(parent_obj, property)
-            ):
-                continue
-            if callable(obj):
-                obj.__doc__ = parent_obj.__doc__
-            elif isinstance(obj, property) and obj.fget is not None:
-                p = property(obj.fget, obj.fset, obj.fdel, parent_obj.__doc__)
-                setattr(cls, attr, p)
-        return cls
-
-    return decorator
-
-
-def try_cast_to_pandas(obj):
-    """
-    Converts obj and all nested objects from modin to pandas if it is possible,
-    otherwise returns obj
-
-    Parameters
-    ----------
-        obj : object,
-            object to convert from modin to pandas
-
-    Returns
-    -------
-        Converted object
-    """
-    if hasattr(obj, "_to_pandas"):
-        return obj._to_pandas()
-    if isinstance(obj, (list, tuple)):
-        return type(obj)([try_cast_to_pandas(o) for o in obj])
-    if isinstance(obj, dict):
-        return {k: try_cast_to_pandas(v) for k, v in obj.items()}
-    if callable(obj):
-        module_hierarchy = getattr(obj, "__module__", "").split(".")
-        fn_name = getattr(obj, "__name__", None)
-        if fn_name and module_hierarchy[0] == "modin":
-            return (
-                getattr(pandas.DataFrame, fn_name, obj)
-                if module_hierarchy[-1] == "dataframe"
-                else getattr(pandas.Series, fn_name, obj)
-            )
-    return obj
-
-
-def wrap_udf_function(func):
-    def wrapper(*args, **kwargs):
-        result = func(*args, **kwargs)
-        # if user accidently returns modin DataFrame or Series
-        # casting it back to pandas to properly process
-        return try_cast_to_pandas(result)
-
-    wrapper.__name__ = func.__name__
-    return wrapper
 
 
 def hashable(obj):
